Add some tests for font-lock faces
* tests.el (forth-parsing-words-font-lock): Add some tests.
(forth-backslash-comment-font-lock): Adjust a little as there are
now more keywords and more correct handling of words that are at
the end-of-buffer.
(forth-assert-face): Test 'face of 'font-lock-face.  We probably
don't care about how it's implemented as long as it looks the
same.
(forth-strip-|-and-→, forth-strip-|-and-¹²): Use remove
instead of delete. That's slightly nicer.

* forth-syntax.el (forth-syntax--state-normal): Call parser
functions directly so that words that are at the end-of-buffer can
be font-locked.