Add test for comment-dwim
That's the command run by M-;.
Not sure what we actually want it to do but the test case is not
entirely unreasonable.

* test/tests.el (forth-comment-dwim): New test.
(forth-paren-comment-font-lock): Changing some of the comment-* variables
also changed the some of font-lock-comment-delimiter-face into
font-lock-comment-face.  Odd.

* forth-mode.el (forth-mode): Use a shy group in the
comment-start-skip regexp, as the non-shy group has some special
meaning (I couldn't quite decipher the docstring) and the test case
seems to work fine with shy groups.
Also, use setq-local as the comment-* variables are not permanently local.