Try to indent (foo)
This is a tricky case because SMIE thinks, depending on
`comment-start-skip` (which indirectly depends on `comment-start`
thru `comment-normalize-vars`), that (foo) is a comment.  But since
(foo) is not actually a comment this leads to an endless recursion.

* forth-mode.el (comment-start-skip, comment-start): Choose them
so that SMIE doesn't run into the endless recursion. (Quite brittle.)

* test/tests.el (forth-indent-\(foo\)): New test.
(forth-backslash-comment-font-lock, forth-comment-dwim): Adjust for
new values of comment-start-skip.

* forth-smiel.el (forth-smie--indentation-rules): For clarity, make
default case explicit.